[#27] test : 리팩토링으로 인한 테스트코드 수정
