+++ src/test/java/com/jinddung2/givemeticon/domain/item/service/ItemVariantServiceTest.java
@@ -3,6 +3,7 @@ package com.jinddung2.givemeticon.domain.item.service;
 import com.jinddung2.givemeticon.domain.item.domain.ItemVariant;
 import com.jinddung2.givemeticon.domain.item.domain.validator.ItemVariantCreateValidator;
 import com.jinddung2.givemeticon.domain.item.dto.request.ItemVariantCreateRequest;
+import com.jinddung2.givemeticon.domain.item.exception.DuplicatedBarcodeException;
 import com.jinddung2.givemeticon.domain.item.mapper.ItemVariantMapper;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -48,15 +49,22 @@ class ItemVariantServiceTest {
         itemVariant.updateItemId(itemId);
         itemVariant.updateSellerId(sellerId);
 
-        Mockito.doNothing().when(itemVariantCreateValidator).validate(itemVariantCreateRequest);
-
         Mockito.when(itemVariantMapper.save(any(ItemVariant.class))).thenReturn(10);
 
         int resultId = itemVariantService.save(itemId, sellerId, itemVariantCreateRequest);
 
-        Mockito.verify(itemVariantCreateValidator).validate(itemVariantCreateRequest);
         Mockito.verify(itemVariantMapper).save(itemVariant);
 
         Assertions.assertEquals(itemVariant.getId(), resultId);
+    }
+
+    @Test
+    @DisplayName("바코드가 중복이라 예외를 던진다")
+    void validate_Fail_Duplicate_Barcode() {
+        Mockito.when(itemVariantMapper.existsByBarcode(itemVariantCreateRequest.barcode()))
+                .thenReturn(true);
+
+        Assertions.assertThrows(DuplicatedBarcodeException.class,
+                () -> itemVariantService.validateDuplicateBarcode(itemVariantCreateRequest.barcode()));
     }
 }
